lighthousemobile/requirements.txt
Django==2.0.2
django-db-mailer==2.3.19
django-filter==1.1.0
django-grappelli==2.11.1
django-model-utils==3.1.1
djangorestframework==3.7.7
Markdown==2.6.11
mysqlclient==1.3.12
pytz==2018.3
wheel==0.24.0
